Align with EU DGC terminology
This is an attempt to align with the Digital Green Certificate (DGC) terminology which the EU regulation is anticipated to be based on.

In essence:

DGC = The European Green Certificate, which is a payload with contents according to the regulation
HCERT = The container of the DGC and potentially other types of health certificates (national, WHO, et. cetera).

Hence, a DGC is encoded within a HCERT, but HCERT MAY be used for payloads defined by other entities.